0.4.0-preview.18 - 2019/03/13 @2018.3

+++ package/Editor/ValidationSuite/Utilities/CIUtils.cs
@@ -1,12 +1,14 @@
 using System;
 using System.IO;
+using System.Text.RegularExpressions;
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace UnityEditor.PackageManager.ValidationSuite
 {
     internal class CIUtils
     {
-        internal const string UpmCIUtilsId = "upm-ci-utils@0.8.2";
+        internal const string UpmCIUtilsId = "upm-ci-utils@0.8.7";
 
         internal static string GetCIUtilsScript()
         {
@@ -39,27 +41,49 @@ namespace UnityEditor.PackageManager.ValidationSuite
             return File.Exists(buildScript) ? buildScript : string.Empty;
         }
 
-        internal static string _Pack(string command, string path, string destinationPath)
+        internal static List<string> _Pack(string command, string path, string destinationPath)
         {
             //Create a copy of the package on the temp folder so that it can be modified
 
             var launcher = new NodeLauncher();
             launcher.WorkingDirectory = path;
             launcher.Script = GetCIUtilsScript();
-            launcher.Args = command + " pack --artifacts-path . --npm-path \"" + NodeLauncher.NpmScriptPath + "\"";
+            launcher.Args = command + " pack --npm-path \"" + NodeLauncher.NpmScriptPath + "\"";
             launcher.Launch();
 
-            var packageName = launcher.OutputLog.ToString().Trim();
+            var outputLines = launcher.OutputLog.ToString().Trim().Split(Environment.NewLine.ToCharArray());
+            List<string> packagePaths = new List<string>();
 
-            //Copy the file to the destinationPath
-            string finalPackagePath = Path.Combine(destinationPath, packageName);
+            Regex packageNameRegex = new Regex(@"^com.*tgz$",
+                RegexOptions.CultureInvariant | RegexOptions.ExplicitCapture);
 
-            if (File.Exists(finalPackagePath))
-                File.Delete(finalPackagePath);
+            foreach (var line in outputLines)
+            {
+                var match = packageNameRegex.Match(line);
+                if (match.Success) {
+                    //Copy the file to the destinationPath
+                    string finalPackagePath = Path.Combine(destinationPath, line);
 
-            File.Move(Path.Combine(path, packageName), finalPackagePath);
+                    if (File.Exists(finalPackagePath))
+                    {
+                        File.Delete(finalPackagePath);
+                    }
 
-            return packageName;
+                    var packagePath = Path.Combine(path, "upm-ci~/packages", line);
+
+                    Debug.LogFormat("Moving {0} to {1}", packagePath, finalPackagePath);
+                    File.Move(packagePath, finalPackagePath);
+                    packagePaths.Add(packagePath);
+                }
+            }
+
+            //See if upm-ci~ exists and remove
+            if (Directory.Exists(Path.Combine(path, "upm-ci~")))
+            {
+                Directory.Delete(Path.Combine(path, "upm-ci~"), true);
+            }
+            
+            return packagePaths;
         }
     }
 }
